blasppConfig: fix optional backends
Relying on the optional "use_openmp" et al. flags to control the
`blasppConfig.cmake` call to `find_dependencies` does not work.

In the case that a user did not find the optionally requested
OpenMP (or CUDA or HIP) backend, the `CMakeLists.txt` scripts
are instructed to continue installing without it. A downstream
script consuming `blasppConfig.cmake` will then fail, because
`find_dependencies()` will have an implicit *required* dependency
on the option that was requested as optional at build time.

Generally, I would recommend to either implement an auto/on/off
logic for those options or to make them only on/off, both are
more standard in the community then the current approach.